artist_performances method moved into UpcomingEvents class
